feat: extract data to viewModel
